+++ packages/secret-numbers/src/schema/Account.ts
@@ -1,99 +1,101 @@
-import * as keypairs from "ripple-keypairs";
+import { deriveAddress, deriveKeypair, generateSeed } from 'ripple-keypairs'
 
-import * as utils from "../utils";
+import {
+  entropyToSecret,
+  parseSecretString,
+  randomSecret,
+  secretToEntropy,
+} from '../utils'
 
 /* Types ==================================================================== */
 
-// eslint-disable-next-line import/no-unused-modules -- it is returned by Account.getKeypair
 export interface Keypair {
-  publicKey: string;
-  privateKey: string;
+  publicKey: string
+  privateKey: string
 }
 
 interface AccountData {
-  familySeed: string;
-  address: string;
-  keypair: Keypair;
+  familySeed: string
+  address: string
+  keypair: Keypair
 }
 
 /* Class ==================================================================== */
 
-export default class Account {
-  private readonly _secret: string[];
+export class Account {
+  private readonly _secret: string[]
   private readonly _account: AccountData = {
-    familySeed: "",
-    address: "",
+    familySeed: '',
+    address: '',
     keypair: {
-      publicKey: "",
-      privateKey: "",
+      publicKey: '',
+      privateKey: '',
     },
-  };
+  }
 
   constructor(secretNumbers?: string[] | string | Uint8Array) {
-    if (typeof secretNumbers === "string") {
-      this._secret = utils.parseSecretString(secretNumbers);
+    if (typeof secretNumbers === 'string') {
+      this._secret = parseSecretString(secretNumbers)
     } else if (Array.isArray(secretNumbers)) {
-      this._secret = secretNumbers;
-    } else if (secretNumbers instanceof Uint8Array) {
-      this._secret = utils.entropyToSecret(secretNumbers);
+      this._secret = secretNumbers
+    } else if (Buffer.isBuffer(secretNumbers)) {
+      this._secret = entropyToSecret(secretNumbers)
     } else {
-      this._secret = utils.randomSecret();
+      this._secret = randomSecret()
     }
 
-    validateLengths(this._secret);
-    this.derive();
+    validateLengths(this._secret)
+    this.derive()
   }
 
   getSecret(): string[] {
-    return this._secret;
+    return this._secret
   }
 
   getSecretString(): string {
-    return this._secret.join(" ");
+    return this._secret.join(' ')
   }
 
   getAddress(): string {
-    return this._account.address;
+    return this._account.address
   }
 
   getFamilySeed(): string {
-    return this._account.familySeed;
+    return this._account.familySeed
   }
 
   getKeypair(): Keypair {
-    return this._account.keypair;
+    return this._account.keypair
   }
 
   toString(): string {
-    return this.getSecretString();
+    return this.getSecretString()
   }
 
   private derive(): void {
     try {
-      const entropy = utils.secretToEntropy(this._secret);
-      this._account.familySeed = keypairs.generateSeed({ entropy });
-      this._account.keypair = keypairs.deriveKeypair(this._account.familySeed);
-      this._account.address = keypairs.deriveAddress(
-        this._account.keypair.publicKey
-      );
+      const entropy = secretToEntropy(this._secret)
+      this._account.familySeed = generateSeed({ entropy })
+      this._account.keypair = deriveKeypair(this._account.familySeed)
+      this._account.address = deriveAddress(this._account.keypair.publicKey)
     } catch (error) {
-      let message = "Unknown Error";
+      let message = 'Unknown Error'
       if (error instanceof Error) {
-        message = error.message;
+        message = error.message
       }
       // we'll proceed, but let's report it
-      throw new Error(message);
+      throw new Error(message)
     }
   }
 }
 
 function validateLengths(secretNumbers: string[]): void {
   if (secretNumbers.length !== 8) {
-    throw new Error("Secret must have 8 numbers");
+    throw new Error('Secret must have 8 numbers')
   }
   secretNumbers.forEach((num) => {
     if (num.length !== 6) {
-      throw new Error("Each secret number must be 6 digits");
+      throw new Error('Each secret number must be 6 digits')
     }
-  });
+  })
 }