+++ packages/ripple-binary-codec/HISTORY.md
@@ -17,6 +17,21 @@
 * Remove `assert` dependency. If you were catching `AssertionError` you need to change to `Error`.
 * Remove `create-hash` in favor of `@noble/hashes`
 
+## 2.0.0 Beta 1 (2023-10-19)
+
+### Breaking Changes
+* Bump typescript to 5.x
+* Remove Node 14 support
+* Remove decimal.js and big-integer. Use `BigNumber` from `bignumber.js` instead of `Decimal` and the native `BigInt` instead of `bigInt`.
+* Remove `assert` dependency. If you were catching `AssertionError` you need to change to `Error`.
+* Remove `create-hash` in favor of `@noble/hashes`
+
+### Changes
+* Update type definitions which causing errors in tests that the code already supported
+  * `makeParser` to accept a `Buffer` in addition to `string`
+  * `SerializedType` constructor allows not passing in a byte array
+  * `Comparable` is now a generic type so that it allows `compareTo` methods to take more that the type itself.
+
 ## 1.10.0 (2023-09-27)
 ### Added
 - Support for the XChainBridge amendment.